+++ cores/epoxy/StdioSerial.h
@@ -19,14 +19,23 @@ class StdioSerial: public Stream {
 
     size_t write(uint8_t c) override;
 
+    // Pull in all other overloaded versions of the write() function from the
+    // Print parent class. This is required because when we override one version
+    // of write() above, C++ performs a static binding to the write() function
+    // in the current class and doesn't bother searching the parent classes for
+    // any other overloaded function that it could bind to. (30 years of C++ and
+    // I still get shot with C++ footguns like this. I have no idea what happens
+    // if the Stream class overloaded the write() function.)
+    using Print::write;
+
     operator bool() { return true; }
 
     int available() override;
 
     int read() override;
-	
+
     int peek() override;
-    
+
   private:
     int bufch;
 };