+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 * Unreleased
+    * Add a `using Print::write` statement in `StdioSerial.h` to
+      pull in all other overloaded `write()` methods from the parent `Print`
+      class.
 * 1.2.1 (2022-01-10)
     * Add `strncasecmp_P()` to `pgmspace.h`. See
       [PR#52](https://github.com/bxparks/EpoxyDuino/pull/52).